+++ features/step_definitions/settings_steps.rb
@@ -47,6 +47,7 @@ end
 Given /^users are blocked for ([0-9]+) minutes after ([0-9]+) failed login attempts$/ do |duration, attempts|
   Setting.brute_force_block_minutes = duration
   Setting.brute_force_block_after_failed_logins = attempts
+end
 
 Given /^we paginate after (\d+) items$/ do |per_page_param|
   Setting.per_page_options = "#{per_page_param}, 50, 100"